+++ src/packet/tcp_header.cpp
@@ -387,7 +387,7 @@ uint16_t TcpHeader::getProtoId() const
     return IPPROTO_TCP;
 }
 
-uint32_t TcpHeader::setHeaderData(const std::vector<unsigned char>& dataPacket)
+uint16_t TcpHeader::setHeaderData(const std::vector<unsigned char>& dataPacket)
 {
     //TODO может добавлять доп.поля (опции)
 }
